appengine/requirements.txt
flask==1.1.1
python-telegram-bot==11.1.0
requests==2.23.0